% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDToGMatrix.r
\name{BED2GMatrix}
\alias{BED2GMatrix}
\title{Create two files needed to run SCORE platform}
\usage{
BED2GMatrix(bfile, ref)
}
\arguments{
\item{bfile}{name of the PLINK files (i.e. if you have "example.bed", 
"example.bim", "example.fam" files as PLINK binary data storage, 
bfile="example")}

\item{ref}{file with a list of DNA variant locations and corresponding 
reference alleles. Could be downloaded from http://dnascore.net -> Tutorial}
}
\description{
This function generates matrix of left singular 
vectors and summary genotype counts. Binary PLINK files are used as an input.
}
