% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTests.R
\name{calcLambda}
\alias{calcLambda}
\title{Calculate Genomic Inflation Factor}
\usage{
calcLambda(pvals, fractionSmallest = 1)
}
\arguments{
\item{pvals}{Numeric vector of p-values}

\item{fractionSmallest}{fracrtion of smallest p-values 
to consider for lambda calculation}
}
\value{
Numeric value of the genomic inflation factor
}
\description{
Calculates the genomic inflation factor (lambda) from a vector of p-values.
If the input p-values are zeros or ones, they are
 replaced with uniform random values
between zero and the first non-zero p-value, and between
the last non-one p-value and one, respectively.
}
\details{
The fraction of smallest p-values to consider
 for lambda calculation can be adjusted
This may be useful in the case of applying to results of
Fisher's exact test, where the p-values are not
ideally uniformly distributed.
}
\examples{
pvals <- c(0.01, 0.05, 0.1, 0.2)
lambda_val <- calcLambda(pvals)
}
