% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{callRateMatrixVCF}
\alias{callRateMatrixVCF}
\title{Scan VCF file for call rate matrix}
\usage{
callRateMatrixVCF(vcf, regions, DP = 10L, GQ = 20L, samples = NULL,
  bannedPositions = NULL, verbose = FALSE)
}
\arguments{
\item{vcf}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}

\item{regions}{the set of regions in format "chr# startPos endPos". For each
region call rate will be calculated and corresponding matrix will be returned.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{samples}{the set of samples to be scanned and returned}

\item{bannedPositions}{the set of positions in format "chr#:#" that 
must be eliminated from consideration.}

\item{verbose}{logical 
metadata file prefix_meta. If not NULL corresponding files will be generated.}
}
\value{
numeric matrix: ratio of non-missing values to all variants in the
region presented in VCF file.
}
\description{
The method parses VCF file and calculate call rate for all specified regions.
}
\examples{
vcf <- "CEU.exon.2010_09.genotypes.vcf.gz"
filepath <- system.file("extdata", vcf, package = "SVDFunctions")

regions <- c("chr1 1108138 36000000", "chr20 33521213 61665700") 
samples <- c("NA06989", "NA10847", "NA11840", "NA12873")

callRateMatrixVCF(filepath, regions, GQ = 0, samples = samples)
}
