% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Constructor for clustering object.}
\usage{
clustering(classification, hierarchy)
}
\arguments{
\item{classification}{named with sample names character vector of 
classes.}

\item{hierarchy}{tree-like structure, a list containing cluster name(leafs) or
two lists with the same structure. See examples.}
}
\description{
Constructor for clustering object.
}
\examples{
samples <- c("C1", "C1", "C2", "C2", "C3", "C4")
names(samples) <- paste0("sample", 1:6)
hier <- list("C1", list(list("C2", "C4"), "C3"))
clustering(samples, hier)
}
