% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeResidual.Cross.R
\name{ComputeResidual.Cross}
\alias{ComputeResidual.Cross}
\title{Computing relative residual vector norm}
\usage{
ComputeResidual.Cross(SampleVector, PreprocessedU)
}
\arguments{
\item{SampleVector}{Vector of genotypes for a sample to be evaluated as control}

\item{PreprocessedU}{\eqn{I - UU^T} matrix calculated with ComputeResidual.preproc}
}
\description{
Computes relative residual vector norm using pre-computed results 
from ComputeResidual.preproc
}
