% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeResidual.preproc.R
\name{computeResidual.preproc}
\alias{computeResidual.preproc}
\title{Intermediate computation of the residual vectors}
\usage{
computeResidual.preproc(referenceU, SV)
}
\arguments{
\item{referenceU}{Matrix of the left singular vectors of cases}

\item{SV}{Number of singular vectors to be used for reconstruction
of the original vector}
}
\description{
Residual vector is estimated as \eqn{(I - UU^T)Z}, where part 
\eqn{I - UU^T} is the same for every 
control sample. Thus, could be precomputed only once for every case 
basis supplied.
}
