% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{dendrogramEstimate}
\alias{dendrogramEstimate}
\title{Estimate dendrogram for case clustering
uses combiM output of clustCombi() function
from Mclust package}
\usage{
dendrogramEstimate(combiM)
}
\arguments{
\item{combiM}{combiM output of clustCombi() function}
}
\description{
Estimate dendrogram for case clustering
uses combiM output of clustCombi() function
from Mclust package
}
