% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{estimateCaseClusters}
\alias{estimateCaseClusters}
\title{Estimate clusters within case data
Clusters are estimated from PC loadings using
Gaussian mixed model fitting from Mclust package
using ellipsoidal, varying volume, shape, and orientation
model ("VVV")}
\usage{
estimateCaseClusters(PCA, plotBIC = FALSE, plotDendrogram = FALSE)
}
\arguments{
\item{PCA}{results of gmatrix principal component analysis}

\item{plot}{logical whether a plot of Bayesian Information 
Criterion vs cluster number should be returned}
}
\description{
Estimate clusters within case data
Clusters are estimated from PC loadings using
Gaussian mixed model fitting from Mclust package
using ellipsoidal, varying volume, shape, and orientation
model ("VVV")
}
