% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{filterGmatrix}
\alias{filterGmatrix}
\title{Filter gmatrix by minimal sample call rate and minimal variant call rate.}
\usage{
filterGmatrix(
  gmatrix,
  imputationResults,
  minVariantCallRate = 0.95,
  minSampleCallRate = 0.95
)
}
\arguments{
\item{gmatrix}{gmatrix with imputed missing values.}

\item{imputationResults}{matrix of the same dimensions as \code{gmatrix}
filled with values TRUE/FALSE indicating which genotypes were imputed.}

\item{minVariantCallRate}{minimal variant call rate.}

\item{minSampleCallRate}{minimal sample call rate.}
}
\description{
Filter gmatrix by minimal sample call rate and minimal variant call rate.
}
