% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{finSwedDataset}
\alias{finSwedDataset}
\title{Finnish & Swedish exome collection dataset preparation pack.}
\format{
An object of class \code{list} of length 3.
}
\usage{
finSwedDataset
}
\description{
A dataset containing all required information that is necessary for
creating an instance that can be submitted to website \url{http://dnascore.net}.
}
\keyword{datasets}
