% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTests.R
\name{genoPhenoVecPval}
\alias{genoPhenoVecPval}
\title{Vectorized Linear Regression P-value}
\usage{
genoPhenoVecPval(case_0, case_1, case_2, control_0, control_1, control_2)
}
\arguments{
\item{case_0}{Number of cases with genotype 0}

\item{case_1}{Number of cases with genotype 1}

\item{case_2}{Number of cases with genotype 2}

\item{control_0}{Number of controls with genotype 0}

\item{control_1}{Number of controls with genotype 1}

\item{control_2}{Number of controls with genotype 2}
}
\value{
Numeric p-value
}
\description{
Vectorized version of \code{genoPhenoLM} returning only the p-value.
}
\examples{
# Vectorized usage:
genoPhenoVecPval(
  case_0 = c(10, 11), case_1 = c(20, 21), case_2 = c(30, 31),
  control_0 = c(15, 16), control_1 = c(25, 26), control_2 = c(35, 36)
)
}
