% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{genotypeMatrixVCF}
\alias{genotypeMatrixVCF}
\title{Scan VCF file for genotypes}
\usage{
genotypeMatrixVCF(vcf, DP = 10L, GQ = 20L, variants = NULL,
  samples = NULL, bannedPositions = NULL, predictMissing = FALSE,
  verbose = FALSE)
}
\arguments{
\item{vcf}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*". This filter is 
applied only for genotype matrix.}

\item{samples}{the set of samples to be scanned and returned}

\item{bannedPositions}{the set of positions in format "chr#:#" that 
must be eliminated from consideration.}

\item{predictMissing}{if TRUE missing values will be replaced with predicted
values.}

\item{verbose}{logical}
}
\description{
The method scans VCF files and returns genotype matrix for specified 
variants after applying neccessary filters.
}
\examples{
vcf <- "CEU.exon.2010_09.genotypes.vcf.gz"
filepath <- system.file("extdata", vcf, package = "SVDFunctions")

samples <- c("NA06989", "NA10847", "NA11840", "NA12873")
genotypeMatrixVCF(filepath, GQ = 0, samples = samples) 
}
