% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{genotypesToCounts}
\alias{genotypesToCounts}
\title{Collect counts statistics from genotype matrix}
\usage{
genotypesToCounts(genotypeMatrix)
}
\arguments{
\item{genotypeMatrix}{matrix containing values 0 for sample with both 
reference alleles, 1 for heterozygous samples, and 2 - with both alternative
alleles. Missing values are allowed}
}
\value{
data.frame with counts statistics
}
\description{
Calculate counts statistics for genotype matrix and return data.frame
}
