% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{gmatrixPCA}
\alias{gmatrixPCA}
\title{Estimate PCA from gmatrix without missing values}
\usage{
gmatrixPCA(gmatrix, components = 30)
}
\arguments{
\item{gmatrix}{Genotype matrix wihout missing values}
}
\description{
PCA is estimated using SVD of the mean centered genotype matrix
PCA coordinates will be needed to perform clustering of the 
case samples to ensure relatively homogenious matching.
}
