% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{gmatrixQC}
\alias{gmatrixQC}
\title{QC gmatrix
performs callrate estimate on samples and variants
checks for Hardy-Weinberg equilibrium and finds consistent
set of variants across all clusters}
\usage{
gmatrixQC(gmatrix, imputationResults, clusters,
  minVariantCallRate = 0.95, minSampleCallRate = 0.95)
}
\description{
QC gmatrix
performs callrate estimate on samples and variants
checks for Hardy-Weinberg equilibrium and finds consistent
set of variants across all clusters
}
