% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelResidEstimate.R
\name{parallelResidEstimate}
\alias{parallelResidEstimate}
\title{Estimation of residual vector norms for all controls.}
\usage{
parallelResidEstimate(
  genotypeMatrix,
  SVDReference,
  caseMeans,
  contMeans,
  SV = 3
)
}
\arguments{
\item{genotypeMatrix}{Genotype matrix.}

\item{SVDReference}{Reference basis of the left singular vectors.}

\item{caseMeans}{numeric vector representing per-variant mean genotype value
of the cases.}

\item{contMeans}{numeric vector representing per-variant mean genotype value
of the controls.}

\item{SV}{number of singular vectors to be used.}
}
\description{
Estimation of residual vector norms for all controls.
}
