% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchingPlot.R
\name{plotMatching}
\alias{plotMatching}
\title{Plot details of a performed matching.}
\usage{
plotMatching(
  popObj,
  variants,
  controlsU,
  meanControl,
  casesGenotypeMatrix,
  controlsGenotypeMatrix,
  matchedControls
)
}
\arguments{
\item{popObj}{a subpopulation from instance read by function readInstanceFromYml.
For an instance object inst an objects like inst$population[[1]] can be passed.}

\item{variants}{list of variants gathered from instance object.}

\item{controlsU}{a numerical matrix carrying vectors of U component from SVD.
It may contain superset of variants provided by variants parameter. 
Rownames should be set to variant signatures.}

\item{meanControl}{mean genotype value of variants in the order of rows of controlsU matrix.}

\item{casesGenotypeMatrix}{case genotype matrix, row names must be set to variant signatures.}

\item{controlsGenotypeMatrix}{control genotype matrix, row names must be set to variant signatures.
colnames must be set to sample names.}

\item{matchedControls}{names of matched controls.}
}
\description{
The function is aimed at plotting the both cohorts (case and control) 
in terms of PCA, the MVN distribution is shown as a set of loadings 
projected into subspace spanned by PC loadings of control cohort.
}
