% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{prepareInstance1}
\alias{prepareInstance1}
\title{prepare instance and write yaml file from QC-ed gmatrix}
\usage{
prepareInstance1(gmatrix, imputationResults, minVariantCallRate = 0.95,
  minSampleCallRate = 0.95, outputFileName, title = "DNAScoreInput")
}
\arguments{
\item{gmatrix}{gmatrix with imputed missing values}

\item{imputationResults}{matrix indicating which genotypes
were imputed}

\item{minVariantCallRate}{minimal variant call rate}

\item{minSampleCallRate}{minimal sample call rate}

\item{outputFileName}{name of the YAML file to output}

\item{title}{title to put as a first line in yaml file
not working so far}
}
\description{
prepare instance and write yaml file from QC-ed gmatrix
}
