% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControls.R
\name{preprocess_controls}
\alias{preprocess_controls}
\title{Performs}
\usage{
preprocess_controls(
  genotypeMatrix,
  SVDReference,
  controlsMean,
  controlsClustering,
  filename
)
}
\description{
Performs
}
