% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTests.R
\name{qqPlot}
\alias{qqPlot}
\title{QQ Plot for P-values}
\usage{
qqPlot(pvals, title = "QQ Plot")
}
\arguments{
\item{pvals}{Numeric vector of p-values}

\item{title}{Title for the plot}
}
\value{
A ggplot object
}
\description{
Generates a QQ plot for the provided p-values.
}
