% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{readInstanceFromYml}
\alias{readInstanceFromYml}
\title{Read instance from YML file.}
\usage{
readInstanceFromYml(filename)
}
\arguments{
\item{filename}{YML file.}
}
\value{
a list of four: title of dataset, version, list of variants, hierarchy and
population description.
}
\description{
Read instance from YML file.
}
