% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{removeCluster}
\alias{removeCluster}
\title{Remove cluster from clusutering object with all samples in it.}
\usage{
removeCluster(clustering, cluster)
}
\arguments{
\item{clustering}{clustering object.}

\item{cluster}{a name of the cluster to delete or its id.}
}
\description{
Remove cluster from clusutering object with all samples in it.
}
