% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMissing.R
\name{ReplaceMissing}
\alias{ReplaceMissing}
\title{Replace of missing values in genotype matrix}
\usage{
ReplaceMissing(GenotypeMatrix)
}
\arguments{
\item{GenotypeMatrix}{Matrix of genotypes (class – Matrix)}
}
\description{
SVD does not tolerated missing values in a matrix. This function replaces 
missing entries in genotype matrix with average genotype for each DNA 
variant rounded to an integer.
}
