% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{sameSizeKMeans}
\alias{sameSizeKMeans}
\title{Function implements clustering algorithm for controls 
based on the same size k-means algorithm.}
\usage{
sameSizeKMeans(gmatrix, k, max_iter = 10000, tol = 1e-04)
}
\arguments{
\item{gmatrix}{Genotype matrix.}

\item{k}{cluster size}

\item{max_iter}{maximum number of iterations}

\item{tol}{tolerance for convergence}
}
\value{
vector of cluster ids for each sample.
}
\description{
Function implements clustering algorithm for controls 
based on the same size k-means algorithm.
}
