% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{sampleNamesVCF}
\alias{sampleNamesVCF}
\title{Scan VCF file for sample names}
\usage{
sampleNamesVCF(vcf)
}
\arguments{
\item{vcf}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}
}
\value{
character vector with sample names
}
\description{
The method scans only header of VCF file and returns a vector with sample 
names
}
\examples{
vcf <- "CEU.exon.2010_09.genotypes.vcf.gz"
filepath <- system.file("extdata", vcf, package = "SVDFunctions")

sampleNamesVCF(filepath)
}
