% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{scanBinaryFile}
\alias{scanBinaryFile}
\title{Scan binary variant file}
\usage{
scanBinaryFile(binaryFile, metafile, samples, variants = NULL,
  regions = NULL, DP = 10, GQ = 20)
}
\arguments{
\item{binaryFile}{the name of binary file}

\item{metafile}{the name of metadata file}

\item{samples}{the set of samples to be analyzed}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*". This filter is 
applied only for genotype matrix.}

\item{regions}{the set of regions in format "chr# startPos endPos". For each
region call rate will be calculated and corresponding matrix will be returned.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}
}
\value{
matrix with three columns:  
number of samples with 
both reference alleles, with  one reference allele and one alternative allele,
and with both alternative ones respectively.
}
\description{
Scan binary file produced by \code{\link{scanVCF}} and collect required 
per variant allele counts statistics after applying quality control filters.
}
