% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{scanVCF}
\alias{scanVCF}
\title{Scan VCF file.}
\usage{
scanVCF(
  vcf,
  DP = 10L,
  GQ = 20L,
  samples = NULL,
  bannedPositions = NULL,
  variants = NULL,
  returnGenotypeMatrix = TRUE,
  predictMissing = FALSE,
  missingRateThreshold = 0.1,
  regions = NULL,
  binaryPathPrefix = NULL,
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{vcf}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{samples}{the set of samples to be scanned and returned.}

\item{bannedPositions}{the set of positions in format "chr#:#" that 
must be eliminated from consideration.}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*".}

\item{returnGenotypeMatrix}{logical: whether or not return genotype matrix.}

\item{predictMissing}{if TRUE missing values will be replaced with predicted
values.}

\item{missingRateThreshold}{the rows of VCF file with missing rate more
than threshold will be filtered out.}

\item{regions}{the set of regions in format "chr# startPos endPos". For each
region call rate will be calculated and corresponding matrix will be returned.}

\item{binaryPathPrefix}{the path prefix for binary file prefix_bin and 
metadata file prefix_meta. If not NULL corresponding files will be generated.}

\item{verbose}{logical}

\item{seed}{integer: seed for random number generator (for reproducible imputation). Default is 42.}
}
\value{
list containing genotype matrix and/or call rate matrix if 
requested. If
\code{predictMissing = TRUE}, the \code{genotype} element is a list with:
\itemize{
  \item \code{genotype}: numeric matrix with missing values replaced by predictions
  \item \code{predicted}: logical matrix indicating which entries were originally missing
  \item \code{loo}: data.frame with per-variant imputation QC based on out-of-bag
    (leave-one-out-like) predictions on observed genotypes (columns: \code{variant},
    \code{n_observed}, \code{n_missing}, \code{oob_mae}, \code{oob_rmse}, \code{rounded_acc})
}
}
\description{
Scan .vcf or .vcf.gz files in matrix and return genotype matrix, call rate
with applied filters. In addition, the method can generate binary and 
metadata files for faster access to genotype data.
}
