% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{select_controls}
\alias{select_controls}
\title{Selection of the optimal set of controls}
\usage{
select_controls(gmatrix, residuals, case_counts, min = 500)
}
\arguments{
\item{gmatrix}{Genotype matrix}

\item{residuals}{Vector of relative residual vector norms for all prospective controls}

\item{case_counts}{Matrix with summary genotype counts from cases}

\item{min}{Minimal size of a control set that is permitted for return}
}
\description{
Calls C++ routine to find optimal set of controls satisfying 
\eqn{\lambda_GC<1.05} or if none exists – will select a set of controls with
 minimal \eqn{\lambda} GC satisfying \eqn{\lambda < 1.3}.
Otherwise no results will be returned. Minimal size of control set 
is \code{min} samples for privacy preservation reasons.
}
