% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{writeYaml}
\alias{writeYaml}
\title{Write yaml file given the list of US and case counts for 
all clusters}
\usage{
writeYaml(clusterResults, outputFileName, collapsing,
  title = "DNAscoreInput")
}
\arguments{
\item{clusterResults}{list of 2 objects - list of US and list of 
case counts list(US, case_counts)}

\item{outputFileName}{name of the yaml file to write to}

\item{collapsing}{matrix representing cluster collapsing
scheme}

\item{title}{dataset name}
}
\description{
Write yaml file given the list of US and case counts for 
all clusters
}
