% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeResidual.preproc.R
\name{ComputeResidual.preproc}
\alias{ComputeResidual.preproc}
\title{Intermediate computation of the residual vectors}
\usage{
ComputeResidual.preproc(ReferenceU, SV)
}
\arguments{
\item{SampleVector}{Number of singular vectors to be used for reconstruction
of the original vector}

\item{PreprocessedU}{Matrix of the left singular vectors of cases}
}
\description{
Residual vector is estimated as \eqn{(I - UU^T)Z}, where part 
\eqn{I - UU^T} is the same for every 
control sample. Thus, could be pre-computed only once for every case 
basis supplied.
}
