% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{estimateCaseClusters}
\alias{estimateCaseClusters}
\title{Estimate clusters within case data.}
\usage{
estimateCaseClusters(
  PCA,
  plotBIC = FALSE,
  plotDendrogram = FALSE,
  minClusters = 1,
  clusters = 20,
  keepSamples = NULL
)
}
\arguments{
\item{PCA}{results of gmatrix principal component analysis.}

\item{plotBIC}{logical whether a plot of Bayesian Information 
Criterion vs cluster number should be returned.}

\item{plotDendrogram}{whether dendrogram should be plotted.}

\item{minClusters}{minimal number of clusters.}

\item{clusters}{maximum number of clusters.}

\item{keelSamples}{names of the samples to keep in results. The feature is
used in order to cluster jointly with another dataset.}
}
\description{
Clusters are estimated from PC loadings using
Gaussian mixed model fitting from Mclust package
using ellipsoidal, varying volume, shape, and orientation
model ("VVV").
}
