% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{estimateCaseClusters}
\alias{estimateCaseClusters}
\title{Estimate clusters within case data
Clusters are estimated from PC loadings using
Gaussian mixed model fitting from Mclust package
using ellipsoidal, varying volume, shape, and orientation
model ("VVV")}
\usage{
estimateCaseClusters(
  PCA,
  plotBIC = FALSE,
  plotDendrogram = FALSE,
  minClusters = 1,
  clusters = 20,
  keepSamples = NULL
)
}
\arguments{
\item{PCA}{results of gmatrix principal component analysis}

\item{plotBIC}{logical whether a plot of Bayesian Information 
Criterion vs cluster number should be returned}

\item{plotDendrogram}{whether dendrogram should be plotted}

\item{clusters}{maximum number of clusters}
}
\description{
Estimate clusters within case data
Clusters are estimated from PC loadings using
Gaussian mixed model fitting from Mclust package
using ellipsoidal, varying volume, shape, and orientation
model ("VVV")
}
