% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{gmatrixPCA}
\alias{gmatrixPCA}
\title{Estimate PCA from gmatrix without missing values.}
\usage{
gmatrixPCA(gmatrix, SVDReference = NULL, referenceMean = NULL, components = 10)
}
\arguments{
\item{gmatrix}{Genotype matrix wihout missing values.}

\item{SVDReference}{a U matrix of SVD decomposition of centered reference
genotype matrix.}

\item{referenceMean}{numeric vector representing per-variant genotype mean 
value of the genotype matrix that were used to construct \code{SVDReference}.}

\item{components}{Number of principal components to be computed.}
}
\description{
PCA is estimated using SVD of the mean centered genotype matrix.
PCA coordinates will be needed to perform clustering of the 
case samples to ensure relatively homogenious matching.
}
