% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R
\name{gmatrixPCA}
\alias{gmatrixPCA}
\title{Estimate PCA from gmatrix without missing values and make 3dplot}
\usage{
gmatrixPCA(gmatrix, clusters = NULL)
}
\arguments{
\item{gmatrix}{Genotype matrix wihout missing values}

\item{clusters}{character vector of cluster assignment for every
sample in gmatrix}
}
\description{
PCA is estimated using SVD of the mean centered genotype matrix
PCA coordinates will be needed to perform clustering of the 
case samples to ensure relatively homogenious matching.
}
