% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mergeCluster}
\alias{mergeCluster}
\title{Merge a cluster with its sibling in a tree}
\usage{
mergeCluster(clustering, cluster)
}
\arguments{
\item{clustering}{clustering object}

\item{cluster}{a name of the cluster or its id}
}
\description{
Merge a cluster with its sibling in a tree
}
