% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelResidEstimate.R
\name{ParallelResidEstimate}
\alias{ParallelResidEstimate}
\title{Estimation of residual vector norms for all controls}
\usage{
ParallelResidEstimate(gmatrix, svdReference, nSV)
}
\arguments{
\item{gmatrix}{Genotype matrix}

\item{svdReference}{Reference basis of the left singular vectors}

\item{nSV}{Number of singular vectors to be used for reconstruction of the 
original vector}
}
\description{
Estimation of residual vector norms for all controls
}
