% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelResidEstimate.R
\name{ParallelResidEstimate}
\alias{ParallelResidEstimate}
\title{Estimation of residual vector norms for all controls}
\usage{
ParallelResidEstimate(genotypeMatrix, SVDReference, nSV)
}
\arguments{
\item{genotypeMatrix}{Genotype matrix}

\item{SVDReference}{Reference basis of the left singular vectors}

\item{nSV}{Number of singular vectors to be used for reconstruction of the 
original vector}
}
\description{
Estimation of residual vector norms for all controls
}
