% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAncestry.R
\name{predictAncestry}
\alias{predictAncestry}
\title{SVD-based ancestry prediction.}
\usage{
predictAncestry(genotypeMatrix, referenceUList, SV, ancestryList)
}
\arguments{
\item{genotypeMatrix}{Vector of genotypes for a sample.}

\item{referenceUList}{List object containing matrices of the left singular
vectors for every ancestry reference.}

\item{SV}{Number of singular vectors to use.}

\item{ancestryList}{Vector of ancestry names.}
}
\description{
Requires several left singular vector bases from every ancestry to be detected. 
1000 genomes data could be used as training set for creation of reference SVD bases. 
Sample genotypes vector is reconstructed from every supplied basis and basis
 with a smallest relative residual vector norm is chosen as sample’s ancestry.
}
