% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_preparation.R, R/prepareInstance.R
\name{prepareInstance}
\alias{prepareInstance}
\title{prepare instance and write yaml file from QC-ed gmatrix}
\usage{
prepareInstance(gmatrix, imputationResults, outputFileName,
  maxVectors = 50, minVariantCallRate = 0.95,
  minSampleCallRate = 0.95, title = "DNAScoreInput")

prepareInstance(gmatrix, imputationResults, outputFileName,
  maxVectors = 50, minVariantCallRate = 0.95,
  minSampleCallRate = 0.95, title = "DNAScoreInput")
}
\arguments{
\item{gmatrix}{gmatrix with imputed missing values}

\item{imputationResults}{matrix indicating which genotypes
were imputed}

\item{outputFileName}{name of the YAML file to output}

\item{maxVectors}{maximum number of principal directions for each cluster
to be computed and written to a file}

\item{minVariantCallRate}{minimal variant call rate}

\item{minSampleCallRate}{minimal sample call rate}

\item{title}{title to put as a first line in yaml file
not working so far}

\item{gmatrix}{gmatrix with imputed missing values}

\item{imputationResults}{matrix indicating which genotypes
were imputed}

\item{minVariantCallRate}{minimal variant call rate}

\item{minSampleCallRate}{minimal sample call rate}

\item{outputFileName}{name of the YAML file to output}

\item{title}{title to put as a first line in yaml file}
}
\description{
prepare instance and write yaml file from QC-ed gmatrix

prepare instance and write yaml file from QC-ed gmatrix
}
