% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{prepareInstance}
\alias{prepareInstance}
\title{Prepare instance and write yaml file from QC-ed gmatrix}
\usage{
prepareInstance(
  gmatrix,
  imputationResults,
  controlsU,
  meanControl,
  outputFileName,
  clusters = NULL,
  title = "DNAScoreInput",
  MAC = 10,
  MAF = 0.01,
  knn_drop = 0.05,
  normalize_drop = 0.05
)
}
\arguments{
\item{gmatrix}{gmatrix with imputed missing values}

\item{imputationResults}{matrix indicating which genotypes
were imputed}

\item{outputFileName}{name of the YAML file to output}

\item{clusters}{clustering object}

\item{title}{title to put as a first line in yaml file}

\item{MAC}{minor allele count used for QC}

\item{MAF}{minor allele frequency used for QC}

\item{maxVectors}{maximum number of principal directions for each cluster
to be computed and written to a file}
}
\description{
Prepare instance and write yaml file from QC-ed gmatrix
}
