% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInstance.R
\name{prepareInstance}
\alias{prepareInstance}
\title{Prepare an instance to be submitted to website dnascore.net}
\usage{
prepareInstance(vcf, dataset, outputPathPrefix = "dnascore",
  vectors = 10, ...)
}
\arguments{
\item{vcf}{vcf file name}

\item{dataset}{one of predefined in this package datasets. List of supported 
datasets: \code{\link{finSwedDataset}}, \code{\link{publicExomesDataset}}.}

\item{outputPathPrefix}{prefix for ouput files}

\item{vectors}{number of vectors to be generated}
}
\description{
The method generates all required files
}
