% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{scanBinaryFile}
\alias{scanBinaryFile}
\title{Scan binary variant file}
\usage{
scanBinaryFile(
  binaryFile,
  metafile,
  samples,
  variants = NULL,
  regions = NULL,
  DP = 10,
  GQ = 20,
  minMAF = 0,
  maxMAF = 1,
  minMAC = 1L,
  maxMAC = .Machine$integer.max,
  reportSingletons = TRUE,
  minCallRate = 0.9
)
}
\arguments{
\item{binaryFile}{the name of binary file}

\item{metafile}{the name of metadata file}

\item{samples}{the set of samples to be analyzed}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*".}

\item{regions}{the set of regions [startPos, endPos] in format 
"chr# startPos endPos". Regions must be non-overlapping.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{minMAF}{numeric minimum minor allele frequency}

\item{maxMAF}{numeric maximum minor allele frequency}

\item{minMAC}{integer minimum minor allele count}

\item{maxMAC}{integer maximum minor allele count}

\item{reportSingletons}{logical if TRUE singletons will be reported}

\item{minCallRate}{numeric minimum call rate}
}
\value{
matrix with three columns:  
number of samples with 
both reference alleles, with  one reference allele and one alternative allele,
and with both alternative ones respectively.
}
\description{
Scan binary file produced by \code{\link{scanVCF}} and collect required 
per variant allele counts statistics after applying quality control filters.
}
