% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{scanVCF}
\alias{scanVCF}
\title{Scan VCF files}
\usage{
scanVCF(filename, DP = 10L, GQ = 20L, samples = NULL,
  bannedPositions = NULL, variants = NULL,
  returnGenotypeMatrix = TRUE, regions = NULL,
  binaryPathPrefix = NULL)
}
\arguments{
\item{filename}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{samples}{the set of samples to be scanned and returned}

\item{bannedPositions}{the set of positions in format "chr#:#" that 
must be eliminated from consideration.}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*".}

\item{returnGenotypeMatrix}{logical: if TRUE genotype matrix will be returned}

\item{regions}{the set of regions in format "chr#:# - chr#:#". For each
region call rate will be calculated and corresponding matrix will be returned.}

\item{binaryPathPrefix}{the path prefix for binary file prefix_bin and 
metadata file prefix_meta. If not NULL corresponding files will be generated.}
}
\value{
list containing genotype matrix and/or call rate matrix if 
requested.
}
\description{
Scan .vcf or .vcf.gz files in matrix and return genotype matrix, call rate
with applied filters. Also can generate binary and metadata files for
faster access to genotype data.
}
