% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{scanVCF}
\alias{scanVCF}
\title{Scan VCF files}
\usage{
scanVCF(vcf, DP = 10L, GQ = 20L, samples = NULL,
  bannedPositions = NULL, variants = NULL,
  returnGenotypeMatrix = TRUE, predictMissing = FALSE,
  excludedPredictorSamples = NULL, missingRateThreshold = 0.1,
  regions = NULL, binaryPathPrefix = NULL, verbose = FALSE)
}
\arguments{
\item{vcf}{the name of file to read, can be plain text VCF file as well
as compressed with gzip or zlib headers.}

\item{DP}{integer: minimum required read depth for position to be considered,
otherwise assumed as missing.}

\item{GQ}{integer: minimum required genotype quality for position to be 
considered, otherwise assumed as missing.}

\item{samples}{the set of samples to be scanned and returned}

\item{bannedPositions}{the set of positions in format "chr#:#" that 
must be eliminated from consideration.}

\item{variants}{the set of variants in format "chr#:# REF ALT"
(i.e. chr23:1532 T GT). In case of deletion ALT must be "*". This filter is 
applied only for genotype matrix.}

\item{returnGenotypeMatrix}{logical: whether or not return genotype matrix}

\item{predictMissing}{if TRUE missing values will be replaced with predicted
values.}

\item{excludedPredictorSamples}{sample names that will not be used for
prediction of missing values. Ignored if predictMissing is FALSE.}

\item{missingRateThreshold}{the rows of VCF file with missing rate more
than threshold will be filtered out.}

\item{regions}{the set of regions in format "chr# startPos endPos". For each
region call rate will be calculated and corresponding matrix will be returned.}

\item{binaryPathPrefix}{the path prefix for binary file prefix_bin and 
metadata file prefix_meta. If not NULL corresponding files will be generated.}

\item{verbose}{logical}
}
\value{
list containing genotype matrix and/or call rate matrix if 
requested.
}
\description{
Scan .vcf or .vcf.gz files in matrix and return genotype matrix, call rate
with applied filters. In addition, the method can generate binary and 
metadata files for faster access to genotype data.
}
