% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControls.R
\name{selectControls}
\alias{selectControls}
\title{Select a set of controls that matches to a set of cases.}
\usage{
selectControls(
  genotypeMatrix,
  originalGenotypeMatrix,
  casesPDs,
  casesMean,
  SVDReference,
  controlsMean,
  caseCounts,
  controlsClustering = NULL,
  minLambda = 0.75,
  softMinLambda = 0.9,
  softMaxLambda = 1.05,
  maxLambda = 1.3,
  min = 500,
  max = 1000,
  step = 50,
  iterations = 1e+05
)
}
\arguments{
\item{genotypeMatrix}{numeric matrix where rows are variants and columns are
samples. The missing values should be imputed prior calling this function.}

\item{originalGenotypeMatrix}{genotype matrix with missing values.}

\item{casesPDs}{numeric matrix where rows are variants and columns are 
principal directions of the data in case dataset.}

\item{casesMean}{numeric vector representing mean per-variant genotype value.}

\item{SVDReference}{reference basis of the left singular vectors.}

\item{controlsMean}{mean value of the reference genotypes.}

\item{caseCounts}{matrix with summary genotype counts from cases.}

\item{controlsClustering}{cluster names for controls.}

\item{minLambda}{minimum possible lambda.}

\item{softMinLambda}{desirable minimum for lambda.}

\item{softMaxLambda}{desirable maximum for lambda.}

\item{maxLambda}{maximum possible lambda.}

\item{min}{minimal number of clusters from control dataset to be returned.}

\item{max}{maximum number of clusters from control dataset to be returned.}

\item{step}{perform matching with the step.}

\item{iterations}{number of simulated annealing iterations per each subset
size.}
}
\description{
Finds an optimal set of controls satisfying 
\eqn{\lambda_GC < softmax_lambda} and \eqn{\lambda_GC > softmin_lambda}
or if none exists – will select a set of controls with
 closest \eqn{\lambda_GC} to the range \eqn{[softmin_lambda; softmax_lambda]} 
 satisfying \eqn{\lambda_GC < max_lambda} and \eqn{\lambda_GC > min_lambda}.
Otherwise no results will be returned. Minimal size of control set 
is \code{min} samples for privacy preservation reasons.
}
