% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControls.R
\name{selectControls}
\alias{selectControls}
\title{Select a set of controls that matches to a set of cases}
\usage{
selectControls(
  genotypeMatrix,
  originalGenotypeMatrix,
  casesPDs,
  casesMean,
  SVDReference,
  controlsMean,
  caseCounts,
  controlsClustering = NULL,
  minLambda = 0.75,
  softMinLambda = 0.9,
  softMaxLambda = 1.05,
  maxLambda = 1.3,
  min = 500,
  max = 1000,
  step = 50,
  iterations = 1e+05
)
}
\arguments{
\item{genotypeMatrix}{Genotype matrix}

\item{originalGenotypeMatrix}{Genotype matrix with no imputation applied}

\item{SVDReference}{Reference basis of the left singular vectors}

\item{caseCounts}{Matrix with summary genotype counts from cases}

\item{controlsClustering}{cluster names for controls}

\item{minLambda}{Minimum possible lambda}

\item{softMinLambda}{Desirable minimum for lambda}

\item{softMaxLambda}{Desirable maximum for lambda}

\item{maxLambda}{Maximum possible lambda}

\item{min}{Minimal size of a control set that is permitted for return}

\item{meanGenotype}{Mean genotype vector}
}
\description{
Finds an optimal set of controls satisfying 
\eqn{\lambda_GC < softmax_lambda} and \eqn{\lambda_GC > softmin_lambda}
or if none exists – will select a set of controls with
 closest \eqn{\lambda_GC} to the range \eqn{[softmin_lambda; softmax_lambda]} 
 satisfying \eqn{\lambda_GC < max_lambda} and \eqn{\lambda_GC > min_lambda}.
Otherwise no results will be returned. Minimal size of control set 
is \code{min} samples for privacy preservation reasons.
}
