% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectControls.R
\name{selectControls}
\alias{selectControls}
\title{Selection of the optimal set of controls}
\usage{
selectControls(genotypeMatrix, SVDReference, caseCounts,
  minLambda = 0.75, softMinLambda = 0.9, softMaxLambda = 1.05,
  maxLambda = 1.3, min = 500, nSV = 5, binSize = 1)
}
\arguments{
\item{genotypeMatrix}{Genotype matrix}

\item{SVDReference}{Reference basis of the left singular vectors}

\item{caseCounts}{Matrix with summary genotype counts from cases}

\item{minLambda}{Minimum possible lambda}

\item{softMinLambda}{Desirable minimum for lambda}

\item{softMaxLambda}{Desirable maximum for lambda}

\item{maxLambda}{Maximum possible lambda}

\item{min}{Minimal size of a control set that is permitted for return}

\item{nSV}{Number of singular vectors to be used for reconstruction of the}

\item{binSize}{sliding window size for optimal lambda search}
}
\description{
Finds an optimal set of controls satisfying 
\eqn{\lambda_GC < softmax_lambda} and \eqn{\lambda_GC > softmin_lambda}
or if none exists – will select a set of controls with
 closest \eqn{\lambda_GC} to the range \eqn{[softmin_lambda; softmax_lambda]} 
 satisfying \eqn{\lambda_GC < max_lambda} and \eqn{\lambda_GC > min_lambda}.
Otherwise no results will be returned. Minimal size of control set 
is \code{min} samples for privacy preservation reasons.
}
