% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{SelectControls}
\alias{SelectControls}
\title{Selection of the optimal set of controls}
\usage{
SelectControls(gmatrix, svdReference, caseCounts, min = 500, nSV = 5)
}
\arguments{
\item{gmatrix}{Genotype matrix}

\item{svdReference}{Reference basis of the left singular vectors}

\item{caseCounts}{Matrix with summary genotype counts from cases}

\item{min}{Minimal size of a control set that is permitted for return}

\item{nSV}{Number of singular vectors to be used for reconstruction of the}
}
\description{
Calls C++ routine to find optimal set of controls satisfying 
\eqn{\lambda_GC<1.05} or if none exists – will select a set of controls with
 minimal \eqn{\lambda} GC satisfying \eqn{\lambda < 1.3}.
Otherwise no results will be returned. Minimal size of control set 
is \code{min} samples for privacy preservation reasons.
}
