% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{selectControlsHier}
\alias{selectControlsHier}
\title{Select a set of controls that populationally matches a set of cases.}
\usage{
selectControlsHier(
  controlGMatrix,
  originalControlGMatrix,
  cases,
  SVDReference,
  controlsMean,
  clusterMergeCoef = 1.1,
  softMinLambda = 0.9,
  softMaxLambda = 1.05,
  ...
)
}
\arguments{
\item{controlGMatrix}{numeric matrix(0 - ref, 1 - het, 2 - both alt).
Intermediate values are allowed, NAs are not. Rows must already be filtered
to \code{cases$variants} and be in the same order.}

\item{originalControlGMatrix}{integer matrix(0 - ref, 1 - het, 2 - both alt)
with missing values allowed. Rows must already match \code{controlGMatrix}
and therefore \code{cases$variants}.}

\item{cases}{result of calling function readInstanceFromYml.}

\item{SVDReference}{reference basis of the left singular vectors.}

\item{controlsMean}{mean value of the reference genotypes.}

\item{clusterMergeCoef}{numeric coefficient of preference of merging clusters.}

\item{softMinLambda}{desirable minimum for lambda.}

\item{softMaxLambda}{desirable maximum for lambda.}

\item{...}{parameters to be passed to selectControls function.}
}
\description{
Select a set of controls that populationally matches a set of cases.
}
