% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{selectControlsHier}
\alias{selectControlsHier}
\title{Select a set of controls that matches a set of cases}
\usage{
selectControlsHier(
  controlGMatrix,
  controlVariants,
  cases,
  imputationMatrix = NULL,
  originalControlGMatrix = NULL,
  clusterMergeCoef = 1.1,
  softMinLambda = 0.9,
  softMaxLambda = 1.05,
  ...
)
}
\arguments{
\item{controlGMatrix}{numeric matrix(0 - ref, 1 - het, 2 - both alt). 
Intermediate values are allowed, NAs are not.}

\item{controlVariants}{character vector of variants(format: "chrN:POS\\tREF\\tALT")}

\item{cases}{result of calling function readInstanceFromYml.}

\item{imputationMatrix}{logical matrix of the same dimentions as 
controlGMatrix. Element is set to TRUE if the corresponding value has been
imputed in controlGMatrix.}

\item{originalControlGMatrix}{integer matrix(0 - ref, 1 - het, 2 - both alt).
Missing values are allowed.}

\item{clusterMergeCoef}{numeric coefficient of preference of merging clusters.}

\item{softMinLambda}{Desirable minimum for lambda}

\item{softMaxLambda}{Desirable maximum for lambda}

\item{...}{parameters to be passed to selectControls function.}
}
\description{
Select a set of controls that matches a set of cases
}
